Added querylists app. Moved code from testcases app. Not tested.
