+++ EdSurvey/home/templates/home.htm
@@ -14,7 +14,7 @@
     <div align="right">[Роль v] [Организация v] [Пользователь v] [выйти]</div>
     <a href="/surveys/">Пройти тест</a> |
     <a href="/questions/">Вопросы</a> |
-    <a href="/testcases/">Тест-кейсы</a> |
+    <a href="/querylists/">Тест-кейсы</a> |
     Задания |
     Расписание |
     <a href="/admin/" target="_blank">Админка</a>
@@ -27,7 +27,7 @@
 <div class="page-footer">
     Пройти тест |
     <a href="/questions/">Вопросы</a> |
-    <a href="/testcases/">Тест-кейсы</a> |
+    <a href="/querylists/">Тест-кейсы</a> |
     Задания |
     Расписание |
     <a href="/admin/" target="_blank">Админка</a>
