+++ EdSurvey/querylists/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class QuerylistsConfig(AppConfig):
+    name = 'querylists'
+    verbose_name = 'Каталог опросных листов'
