+++ EdSurvey/querylists/urls.py
@@ -1,7 +1,8 @@
 from django.conf.urls import url
+
 from . import views
+
 
 urlpatterns = [
     url(r'^$', views.index),
-    # url(r'^answersbyquestion/(?P<questionid>\d+)$', views.answers_by_question),
 ]
