--- EdSurvey/testcases/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class TestcasesConfig(AppConfig):
-    name = 'testcases'
-    verbose_name = 'Каталог тест-кейсов'
