--- EdSurvey/testcases/templates/testcases.html
@@ -1,18 +1,0 @@
-{% extends 'home.htm' %}
-{% block mainblock %}
-    <h2>Тест-кейсы</h2>
-    <table border="2">
-        <tr>
-            <th>№</th>
-            <th>Название</th>
-            <th>Описание</th>
-        </tr>
-{% for testcase in testcases %}
-        <tr>
-            <td>{{testcase.id}}</td>
-            <td>{{testcase.name}}</td>
-            <td>{{testcase.description}}</td>
-        </tr>
-{% endfor %}
-    </table>
-{% endblock %}
